+++ packages/client/src/utils/contract.ts
@@ -1,3 +1,4 @@
+import { initEventGateway } from '@streamr/utils'
 import { Contract, ContractReceipt, ContractTransaction } from '@ethersproject/contracts'
 import EventEmitter from 'eventemitter3'
 import { NameDirectory } from '@streamr/network-node'
@@ -5,6 +6,7 @@ import pLimit from 'p-limit'
 import { LoggerFactory } from './LoggerFactory'
 import { tryInSequence } from './promises'
 import shuffle from 'lodash/shuffle'
+import { StreamrClientEventEmitter, InternalEvents, StreamrClientEvents } from '../events'
 
 export interface ContractEvent {
     onMethodExecute: (methodName: string) => void
@@ -145,5 +147,48 @@ export const queryAllReadonlyContracts = <T, C>(
         shuffle(contracts).map((contract: C) => {
             return () => call(contract)
         })
+    )
+}
+
+export const initContractEventGateway = <
+    TSourcePayloads extends any[],
+    TSourceName extends string,
+    TTargetName extends keyof (StreamrClientEvents & InternalEvents)
+>(opts: {
+    sourceName: TSourceName
+    targetName: TTargetName
+    sourceEmitter: {
+        on: (name: TSourceName, listener: (...args: TSourcePayloads) => void) => void
+        off: (name: TSourceName, listener: (...args: TSourcePayloads) => void) => void
+    }
+    targetEmitter: StreamrClientEventEmitter
+    transformation: (...args: TSourcePayloads) => Parameters<(StreamrClientEvents & InternalEvents)[TTargetName]>[0]
+    loggerFactory: LoggerFactory
+}): void => {
+    const logger = opts.loggerFactory.createLogger(module)
+    type Listener = (...args: TSourcePayloads) => void
+    initEventGateway(
+        opts.targetName,
+        (emit: (payload: Parameters<(StreamrClientEvents & InternalEvents)[TTargetName]>[0]) => void) => {
+            const listener = (...args: TSourcePayloads) => {
+                let targetEvent
+                try {
+                    targetEvent = opts.transformation(...args)
+                } catch (err) {
+                    logger.debug('Skip emit event', {
+                        eventName: opts.targetName,
+                        reason: err?.message
+                    })
+                    return
+                }
+                emit(targetEvent)
+            }
+            opts.sourceEmitter.on(opts.sourceName, listener)
+            return listener
+        },
+        (listener: Listener) => {
+            opts.sourceEmitter.off(opts.sourceName, listener)
+        },
+        opts.targetEmitter
     )
 }